# swagger.yml
openapi: 3.0.3
info:
  title: go-crypto-arb API
  version: v2.1.0
  description: |
    REST API for go-crypto-arb, a monitoring-only arbitrage detection system.

    The service exposes latest in-memory market data, order book summaries,
    arbitrage estimates, alerts, provider health, an aggregate TUI snapshot,
    and Prometheus metrics.

    Supported crypto exchange platform identities are `okx`, `bybit`,
    `binance`, `kraken`, `coinbase`, `gateio`, and `bitget`. OKX, Bybit,
    Coinbase, Gate.io, and Bitget are disabled by default in configuration and
    currently use spot public REST market data.

    All `/api/v1/*` endpoints require the `X-API-Key` header. `/health` and
    `/metrics` are public; `/metrics` is only registered when Prometheus metrics
    are enabled in configuration.

    This application is read-only. It does not place orders, execute trades,
    automate buy/sell actions, or submit IBKR orders.
servers:
  - url: http://localhost:8080
    description: Default local backend
...
paths:
...
  /api/v1/prices:
    get:
      tags:
        - Market Data
      summary: Get latest tickers and funding rates
      operationId: getPrices
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Latest spot tickers, futures tickers, and funding rates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/order-books:
    get:
      tags:
        - Market Data
      summary: Get order book summaries or filtered full order books
      description: |
        Without query filters, returns compact `OrderBookSummary` records.
        When any supported filter is supplied, returns full matching `OrderBook`
        records.
      operationId: getOrderBooks
      security:
        - ApiKeyAuth: []
      parameters:
        - name: provider
          in: query
          description: Lower-case provider identity such as `okx`, `bybit`, `binance`, `kraken`, `coinbase`, `gateio`, `bitget`, or `ibkr`.
          schema:
            type: string
            enum:
              - okx
              - bybit
              - binance
              - kraken
              - coinbase
              - gateio
              - bitget
              - ibkr
          example: binance
        - name: exchange
          in: query
          description: Exchange display name such as `Binance` or `Kraken`.
          schema:
            type: string
          example: Binance
        - name: symbol
          in: query
          description: Canonical symbol.
          schema:
            type: string
          example: BTC/USDT
        - name: market
          in: query
          description: Market type.
          schema:
            $ref: "#/components/schemas/MarketType"
          example: spot
      responses:
        "200":
          description: Compact summaries by default, full order books when filtered.
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/OrderBookSummary"
                  - type: array
                    items:
                      $ref: "#/components/schemas/OrderBook"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/providers:
    get:
      tags:
        - Providers
      summary: Get configured providers
      operationId: getProviders
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Configured providers sorted by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/arbitrage/triangular:
    get:
      tags:
        - Arbitrage
      summary: Get triangular arbitrage estimates
      operationId: getTriangularArbitrage
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Depth-aware triangular arbitrage estimates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TriangularOpportunityV2"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/arbitrage/cross-exchange:
    get:
      tags:
        - Arbitrage
      summary: Get cross-exchange arbitrage estimates
      operationId: getCrossExchangeArbitrage
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Depth-aware cross-exchange arbitrage estimates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrossExchangeOpportunityV2"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/arbitrage/spot-futures:
    get:
      tags:
        - Arbitrage
      summary: Get spot-futures basis estimates
      operationId: getSpotFuturesArbitrage
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Depth-aware crypto spot-futures basis estimates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpotFuturesOpportunityV2"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/ibkr/instruments:
    get:
      tags:
        - IBKR
      summary: Get configured IBKR instruments
      description: Returns IBKR instrument market info derived from configured instrument universes. Live data support is partial.
      operationId: getIBKRInstruments
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Configured IBKR instruments.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/ibkr/fx-triangular:
    get:
      tags:
        - IBKR
        - Arbitrage
      summary: Get IBKR FX triangular estimates
      description: Strategy endpoint returning results when configured IBKR FX market data is available.
      operationId: getIBKRFXTriangular
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: IBKR FX triangular estimates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TriangularOpportunityV2"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/ibkr/crypto-futures-basis:
    get:
      tags:
        - IBKR
        - Arbitrage
      summary: Get crypto spot vs IBKR futures basis estimates
      description: Strategy endpoint returning results when relevant crypto spot and IBKR futures data is available.
      operationId: getIBKRCryptoFuturesBasis
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Crypto spot vs IBKR futures basis estimates.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrokerFuturesBasisOpportunity"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/signals/related-assets:
    get:
      tags:
        - Signals
      summary: Get related asset signals
      operationId: getRelatedAssetSignals
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Related asset signal groups.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelatedAssetGroupSignal"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/alerts:
    get:
      tags:
        - Alerts
      summary: Get latest alerts
      operationId: getAlerts
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Latest in-memory alerts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/snapshot:
    get:
      tags:
        - Snapshot
      summary: Get aggregate backend snapshot
      description: Returns the complete latest-state snapshot consumed by the TUI.
      operationId: getSnapshot
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Aggregate latest-state snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    Unauthorized:
      description: Missing, invalid, or unconfigured API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    MarketType:
      type: string
      enum:
        - spot
        - futures
    PricesResponse:
      type: object
      properties:
        prices:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/Ticker"
        futures_prices:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/Ticker"
        funding_rates:
          type: array
          items:
            $ref: "#/components/schemas/FundingRate"
      required:
        - prices
        - futures_prices
        - funding_rates
...
    OrderBook:
      type: object
      properties:
        provider:
          type: string
        exchange:
          type: string
        broker:
          type: string
        symbol:
          type: string
        instrument_id:
          type: string
        base_asset:
          type: string
        quote_asset:
          type: string
        market_type:
          $ref: "#/components/schemas/MarketType"
        asset_class:
          type: string
        bids:
          type: array
          items:
            $ref: "#/components/schemas/OrderBookLevel"
        asks:
          type: array
          items:
            $ref: "#/components/schemas/OrderBookLevel"
        updated_at:
          type: string
          format: date-time
        limited_depth:
          type: boolean
      required:
        - provider
        - exchange
        - symbol
        - base_asset
        - quote_asset
        - market_type
        - asset_class
        - bids
        - asks
        - updated_at
    OrderBookSummary:
      type: object
      properties:
        provider:
          type: string
        exchange:
          type: string
        broker:
          type: string
        symbol:
          type: string
        market_type:
          $ref: "#/components/schemas/MarketType"
        best_bid:
          type: string
        best_ask:
          type: string
        bid_levels:
          type: integer
          format: int32
        ask_levels:
          type: integer
          format: int32
        age_seconds:
          type: string
        updated_at:
          type: string
          format: date-time
      required:
        - provider
        - exchange
        - symbol
        - market_type
        - best_bid
        - best_ask
        - bid_levels
        - ask_levels
        - age_seconds
        - updated_at
...
    ProviderResponse:
      type: object
      properties:
        name:
          type: string
          description: Provider identity. Supported crypto exchange values are `okx`, `bybit`, `binance`, `kraken`, `coinbase`, `gateio`, and `bitget`; broker support currently includes `ibkr`.
          example: okx
        type:
          type: string
          enum:
            - crypto_exchange
            - broker
        enabled:
          type: boolean
        config:
          $ref: "#/components/schemas/ProviderConfig"
      required:
        - name
        - type
        - enabled
        - config
...
    TriangularOpportunityV2:
      type: object
      properties:
        provider:
          type: string
        exchange:
          type: string
        strategy_title:
          type: string
        asset_class:
          type: string
        cycle:
          type: array
          items:
            type: string
        start_asset:
          type: string
        start_amount:
          $ref: "#/components/schemas/Decimal"
        end_amount:
          $ref: "#/components/schemas/Decimal"
        net_profit_percent:
          $ref: "#/components/schemas/Decimal"
        complete_fill:
          type: boolean
        worst_leg:
          type: string
        max_slippage_percent:
          $ref: "#/components/schemas/Decimal"
        legs:
          type: array
          items:
            $ref: "#/components/schemas/LegSimulation"
        status:
          type: string
        updated_at:
          type: string
          format: date-time
      required:
        - provider
        - exchange
        - strategy_title
        - cycle
        - start_asset
        - start_amount
        - end_amount
        - net_profit_percent
        - complete_fill
        - worst_leg
        - max_slippage_percent
        - legs
        - status
        - updated_at
...
    RelatedAssetGroupSignal:
      type: object
      properties:
        group:
          type: string
        assets:
          type: array
          items:
            $ref: "#/components/schemas/RelatedAssetSignal"
        group_average:
          $ref: "#/components/schemas/Decimal"
        calculated_at:
          type: string
          format: date-time
      required:
        - group
        - assets
        - group_average
        - calculated_at
...